+++ day24/solution_test.go
@@ -10,13 +10,13 @@ import (
 func TestPart1Example(t *testing.T) {
 	lines := util.FileAsLines("ex1")
 	result := part1(parse(lines))
-	assert.Equal(t, int16(18), result)
+	assert.Equal(t, 18, result)
 }
 
 func TestPart2Example(t *testing.T) {
 	lines := util.FileAsLines("ex1")
 	result := part2(parse(lines))
-	assert.Equal(t, int16(54), result)
+	assert.Equal(t, 54, result)
 }
 
 func BenchmarkPart1(b *testing.B) {
